ml-service/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# ML libraries
scikit-learn==1.4.0
xgboost==2.0.3
prophet==1.1.5

# Data processing
pandas==2.2.0
numpy==1.26.3

# Model serialization
joblib==1.3.2

# HTTP client (for testing)
httpx==0.26.0

# Environment variables
python-dotenv==1.0.0
